pyproject.toml
[tool.poetry]
name = "xaelai"
version = "0.1.0"
description = ""
authors = ["Alex Davies <traverse.da@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
phidata = "v2.4.0"
streamlit = "^1.38.0"
bs4 = "^0.0.2"
psycopg = {extras = ["binary"], version = "^3.2.1"}
sqlalchemy = "^2.0.34"
ollama = "^0.3.2"
pgvector = "^0.3.2"
duckduckgo-search = "^6.2.11"
arxiv = "^2.1.3"
pypdf = "^4.3.1"
sympy = "^1.13.2"
pint = "^0.24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
